% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xml_legend.R
\name{create_xml_legend}
\alias{create_xml_legend}
\title{Create .tif.aux.xml legend for soilgrids layer}
\usage{
create_xml_legend(unique_values, output_dir, file_name)
}
\arguments{
\item{unique_values}{int vector. with unique values in soilgrids layer.}

\item{output_dir}{character. Output directory}

\item{file_name}{character. Name of the output file without .tif.aux.xml extension.}
}
\description{
Create .tif.aux.xml legend for soilgrids layer
}
\keyword{internal}
