% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{degrees}
\alias{degrees}
\title{Radians to degrees}
\usage{
degrees(radian)
}
\arguments{
\item{radian}{Angle in radians and decimal fraction.}
}
\value{
Angle in degrees.
}
\description{
Accessory function to transform radians into degrees.
}
\examples{

\dontrun{
degrees(seq(0,2*pi,pi/2))
}

}
\seealso{
\code{\link{radians}}
}
\keyword{internal}
