% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa_current.R
\name{get_chelsa_current}
\alias{get_chelsa_current}
\title{Create multilayer Tiff file with 107 variables from chelsa-climate.org}
\usage{
get_chelsa_current(
  extent,
  EPSG,
  destination,
  resolution = 1000,
  rm_download = FALSE
)
}
\arguments{
\item{extent}{character. First output of \code{transform_shp_coutry_extend}}

\item{EPSG}{int. to consider for this country/area.}

\item{destination}{character. absolute path where to download files like \code{here()} output.}

\item{resolution}{int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.}

\item{rm_download}{boolean. If TRUE remove download files and folders. Keep only environ.tif in \code{data_raw} folder, default is FALSE.}
}
\value{
character. absolute path to climate_chelsa.tif.
}
\description{
Gives the average values on the data recovered between 1981 and 2010.
Monthly variables are average temperatures, min temperatures, max temperatures, precipitation, potential evapotranspiration (with Penman formula and with Thornthwaite formula),
and total cloud cover.
Others variables are climatic water deficit (with Penman and Thornthwaite), number of dry month (with Penman and Thornthwaite) and 19 bio variables (more information in chelsa documentation).
}
\details{
\code{resolution} need to be carefully choosen because if Tiff file is too big, R can crash.

Unit of each climatic variable :\tabular{ll}{
   Name \tab Unit \cr
   Temperature average (tas) \tab °C x 10 \cr
   Temperature min (tasmin) \tab °C x 10 \cr
   Temperature max (tasmax) \tab °C x 10 \cr
   Precipitation \tab kg.m^{-1} \cr
   Cloud area fraction (clt) \tab \% \cr
   PET Penman \tab kg.m^{-1} \cr
   PET Thornthwaite \tab kg.m^{-1} \cr
   Climatic Water Deficit (Penman) \tab kg.m^{-1} \cr
   Climatic Water Deficit (Thornthwaite) \tab kg.m^{-1} \cr
   Number of dry month (Penman) \tab month \cr
   Number of dry month (Thornthwaite) \tab month \cr
   bio 1 \tab °C x 10 \cr
   bio 2 \tab °C x 10 \cr
   bio 3 \tab °C x 10 \cr
   bio 4 \tab °C x 10 \cr
   bio 5 \tab °C x 10 \cr
   bio 6 \tab °C x 10 \cr
   bio 7 \tab °C x 10 \cr
   bio 8 \tab °C x 10 \cr
   bio 9 \tab °C x 10 \cr
   bio 10 \tab °C x 10 \cr
   bio 11 \tab °C x 10 \cr
   bio 12 \tab kg.m^{-2}.year^{-1} \cr
   bio 13 \tab kg.m^{-2}.month^{-1} \cr
   bio 14 \tab kg.m^{-2}.month^{-1} \cr
   bio 15 \tab kg.m^{-2} \cr
   bio 16 \tab kg.m^{-2}.month^{-1} \cr
   bio 17 \tab kg.m^{-2}.month^{-1} \cr
   bio 18 \tab kg.m^{-2}.month^{-1} \cr
   bio 19 \tab kg.m^{-2}.month^{-1} \cr
}
}
