% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa_future.R
\name{get_chelsa_future}
\alias{get_chelsa_future}
\title{Create several multilayer Tiff files with 81 variables from chelsa-climate.org with future climat variables.}
\usage{
get_chelsa_future(
  extent_latlon,
  extent_proj,
  EPSG,
  destination,
  resolution = 1000,
  phase = "2071-2100",
  GCM = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL"),
  SSP = 585,
  rm_download = FALSE
)
}
\arguments{
\item{extent_latlon}{vector. First output of \code{get_aoi_extent()} function.}

\item{extent_proj}{vector. Second output of \code{get_aoi_extent()} function.}

\item{EPSG}{int. to consider for this country/area.}

\item{destination}{character. absolute path where to download files like \code{here()} output.}

\item{resolution}{int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.}

\item{phase}{character. Must be in c("2041-2070", "2071-2100") match to years to download, default is "2071-2100".}

\item{GCM}{character.  Global Climatic Models considered. Must be in c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL"). By default climate simulated by the five GCMs are downloaded.}

\item{SSP}{character or int. Scenario specifier must be in c(126, 370, 585). Corresponding to the the Shared Socio-economic Pathways you want to download, default is 585 for SSP5-RCP8.5 climate as simulated by the GCMs.}

\item{rm_download}{boolean. If TRUE remove download files and folders. Keep only future_chelsa.tif in \code{data_raw} folder, default is FALSE.}
}
\value{
character. absolute path to future_chelsa.tif.
}
\description{
Gives predictions on future values for the choosen phase. Prediction are from multiple models (GFDL-ESM4, IPSL-CM6A-LR, MPI-ESM1-2-HR, MRI-ESM2-0, UKESM1-0-LL).
Creates folder for each model and a folder with mean of this five models.
Monthly variables are average temperatures, min temperatures, max temperatures, precipitation, potential evapotranspiration with Thornthwaite formula.
Others variables are climatic water deficit with Thornthwaite, number of dry month with Thornthwaite and 19 bio variables (more information in chelsa documentation).
}
\details{
\code{resolution} need to be carefully choosen because if Tiff file is too big, R can crash.

Unit of each climatic variable :\tabular{ll}{
   Name \tab Unit \cr
   Temperature average (tas) \tab °C x 10 \cr
   Temperature min (tasmin) \tab °C x 10 \cr
   Temperature max (tasmax) \tab °C x 10 \cr
   Precipitation \tab kg.m^{-2} \cr
   PET Thornthwaite \tab kg.m^{-2} \cr
   Climatic Water Deficit (Thornthwaite) \tab kg.m^{-2} \cr
   Number of dry month (Thornthwaite) \tab month \cr
   bio 1 \tab °C x 10 \cr
   bio 2 \tab °C x 10 \cr
   bio 3 \tab °C x 10 \cr
   bio 4 \tab °C x 10 \cr
   bio 5 \tab °C x 10 \cr
   bio 6 \tab °C x 10 \cr
   bio 7 \tab °C x 10 \cr
   bio 8 \tab °C x 10 \cr
   bio 9 \tab °C x 10 \cr
   bio 10 \tab °C x 10 \cr
   bio 11 \tab °C x 10 \cr
   bio 12 \tab kg.m^{-2}.year^{-1} \cr
   bio 13 \tab kg.m^{-2}.month^{-1} \cr
   bio 14 \tab kg.m^{-2}.month^{-1} \cr
   bio 15 \tab kg.m^{-2} \cr
   bio 16 \tab kg.m^{-2}.month^{-1} \cr
   bio 17 \tab kg.m^{-2}.month^{-1} \cr
   bio 18 \tab kg.m^{-2}.month^{-1} \cr
   bio 19 \tab kg.m^{-2}.month^{-1} \cr
}
}
