% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa_variables.R
\name{get_chelsa_variables}
\alias{get_chelsa_variables}
\title{Download and crop tif files from chelsa-climate.org
Cropping using country name or shapefile or extent.}
\usage{
get_chelsa_variables(extent, EPSG)
}
\arguments{
\item{extent}{character. output of \code{transform_shp_coutry_extend}}

\item{EPSG}{int. to consider for this country/area.}
}
\description{
Download and crop tif files from chelsa-climate.org
Cropping using country name or shapefile or extent.
}
