% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_variables.R
\name{get_env_variables}
\alias{get_env_variables}
\title{Get variables from many sites.}
\usage{
get_env_variables(extent_latlon, extent, EPSG, country_name)
}
\arguments{
\item{extent_latlon}{int vector. in this order c(lon_min, lat_min, lon_max, lat_max).}

\item{extent}{character. First output of \code{transform_shp_country_extent} function.}

\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{character. country name (in english) to considers.}
}
\description{
Get variables from many sites.
}
