% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aoi_extent.R
\name{get_extents_from_e_latlon}
\alias{get_extents_from_e_latlon}
\title{Get extents from e_latlong.}
\usage{
get_extents_from_e_latlon(e_latlon_orig, EPSG_proj, resol)
}
\arguments{
\item{EPSG_proj}{int. EPSG code.}

\item{resol}{num. Resolution used to align the extent.}

\item{e_latlon}{num vector. Named vector of length 4, in this order c(lonmin,
latmin, lonmax, latmax).}
}
\value{
list. With projected extent and latlon extent.
}
\description{
Get extents from e_latlong.
}
