% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodata_dtype_chelsa.R, R/test.R
\name{get_nodata_dtype_chelsa}
\alias{get_nodata_dtype_chelsa}
\title{Get nodata value and data type for Chelsa variables.}
\usage{
get_nodata_dtype_chelsa(nmonths = 1, biovar = TRUE, verbose = TRUE)

get_nodata_dtype_chelsa(nmonths = 1, biovar = TRUE, verbose = TRUE)
}
\arguments{
\item{nmonths}{numeric. Number of months to consider. Must be
between 1 and 12. Default to 1.}

\item{biovar}{boolean. If TRUE (the default), consider the 19
bioclimatic variables in addition to monthly variables.}

\item{verbose}{boolean. If FALSE, do not print message. Default to
TRUE.}
}
\value{
dataframe. The dataframe includes nodata value and data
type for each variable.

dataframe. The dataframe includes nodata value and data
type for each variable.
}
\description{
Ues gdalinfo and regular expression to extract nodata values and
data types for Chelsa climatic variables. The following monthly
climatic variables are considered: clt, pet, pr, tas, tasmax,
tasmin.

Ues gdalinfo and regular expression to extract nodata values and
data types for Chelsa climatic variables. The following monthly
climatic variables are considered: clt, pet, pr, tas, tasmax,
tasmin.
}
\examples{
\dontrun{
df <- get_nodata_dtype_chelsa(nmonths=1, biovar=TRUE, verbose=TRUE)
print(df)
}

}
\author{
Ghislain Vieilledent \href{mailto:ghislain.vieilledent@cirad.fr}{ghislain.vieilledent@cirad.fr}
}
