% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_var.R
\name{get_osm_var}
\alias{get_osm_var}
\title{Create multilayer Tiff file with 3 environmental variables}
\usage{
get_osm_var(
  extent_latlon,
  extent_proj,
  EPSG,
  country_name,
  destination,
  resol = 1000,
  rm_download = FALSE
)
}
\arguments{
\item{extent_latlon}{vector. First output of \code{get_aoi_extent()}
function.}

\item{extent_proj}{vector. Second output of \code{get_aoi_extent()}
function.}

\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{character. country name (in English) which be
use to collect protected areas. This country must be available in
\verb{https://www.protectedplanet.net/en/thematic-areas/wdpa?tab=WDPA}.}

\item{destination}{character. absolute path where to download files
like \code{here()} output.}

\item{resol}{int. Resolution. If in meters, recommended resolutions
are 250m, 500m, 1km, 2km or 5km. The resolution needs to be
carefully chosen. If set too small (e.g. < 250m), raster file
will be too big to fit in memory and R will crash. Default is
1km.}

\item{rm_download}{boolean. If TRUE remove download files and
folders. Keep only environ.tif in \code{data_raw} folder, default is
FALSE.}
}
\value{
character. Absolute path to \code{environ.tif} file.
}
\description{
Variables are distance to roads, distance to cities and town, distance to
rivers and waterbodies.
}
\details{
environ.tif.aux.xml is an extention of environ.tif, it
allows to classify soilgrid variable with QGIS with
RasterAttributeTable extension. Nevertheless it's cause problems
to open it with \code{stars} package but not with \code{terra}. If you have
any problems to open environ.tif, you can remove
environ.tif.aux.xml. This solve all accessibility problems with
\code{stars} and \code{terra} packages.

Unit of each environ variable :\tabular{ll}{
   Name \tab Unit \cr
   Distance road \tab m \cr
   Distance place \tab m \cr
   Distance watering place \tab m \cr
}
}
