% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Create a multilayer Tiff file with environ & climat variables.}
\usage{
merge_files(environ_path, climate_path, destination)
}
\arguments{
\item{environ_path}{character. Absolute path to \code{environ.tif} file, output of \code{get_env_variables}.}

\item{climate_path}{character. Absolute path to \code{current_chelsa.tif} file, output of \code{get_chelsa_variables}.}

\item{destination}{character. Absolute path for the output directory.}
}
\value{
character. Absolute path to \code{gecevar.tif}.
}
\description{
Use output of \code{get_env_variables} and \code{get_chelsa_current} to create an unique file with more than 100 variables.
Can be use with files created by \code{get_chelsa_futur}.
}
\details{
gecevar.tif.aux.xml is an extention of gecevar.tif, it allows to classify soilgrid variable with QGIS with RasterAttributeTable extension.
Nevertheless it's cause problems to open it with \code{stars} package but not with \code{terra}. If you have any problems to open gecevar.tif, you can remove gecevar.tif.aux.xml.
This solve all accessibility problems with \code{stars} and \code{terra} packages.
}
