% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodata_chelsa}
\alias{nodata_chelsa}
\title{NoData value and data type for Chelsa variables.}
\format{
\subsection{\code{nodata_chelsa}}{

A data frame with 25 rows and 5 variables:
\describe{
\item{var}{Climatic variable}
\item{month}{Month of the year}
\item{dtype}{Data type}
\item{nodata}{NoData value}
}
}
}
\source{
\url{https://chelsa-climate.org/}
}
\usage{
nodata_chelsa
}
\description{
A dataset with the nodata value and data type for each climatic
variable provided in Chelsa. The variables are as follows:
}
\keyword{datasets}
