% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{radians}
\alias{radians}
\title{Degrees to radians}
\usage{
radians(degree)
}
\arguments{
\item{degree}{Angle in degrees and decimal fraction.}
}
\value{
Angle in radians.
}
\description{
Accessory function to transform degrees into radians.
}
\examples{

radians(seq(0,360,90))

}
\seealso{
\code{\link{degrees}}
}
