% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_shp_country_extent.R
\name{transform_shp_country_extent}
\alias{transform_shp_country_extent}
\title{Create extent & extent in latitude and longitude}
\usage{
transform_shp_country_extent(
  EPSG,
  country_name = NULL,
  shapefile_path = NULL,
  extent_project = NULL,
  extent_latlon = NULL,
  rm_download = TRUE
)
}
\arguments{
\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{character. English country name, default is NULL.}

\item{shapefile_path}{character. Path to a .shp file, default is NULL.}

\item{extent_project}{int vector. length 4, with this order c(xmin, ymin, xmax, ymax), default is NULL.}

\item{extent_latlon}{int vector. length 4, with this order c(lonmin, latmin, lonmax, latmax), default is NULL.}

\item{rm_download}{boolean. If TRUE remove downloaded shapefile to get area extent, default is TRUE.}
}
\value{
character vector. with extent of the area in one character, other are latlon coord of area and the path to the downloaded shapefile, if rm_download=FALSE.
}
\description{
With a country name, a shapefile or an extent and an EPSG value,
give the extent corresponding to the coordinates system (EPSG) specified and the extent with default EPSG also known as latitude and longitude coordinates.
}
