% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xml_legend.R
\name{create_xml_legend}
\alias{create_xml_legend}
\title{Create .tif.aux.xml legend for soilgrids layer}
\usage{
create_xml_legend(unique_values, destination, name_file)
}
\arguments{
\item{unique_values}{int vector. with unique values in soilgrids layer.}

\item{destination}{character. absolute path where create .tif.aux.xml file.}

\item{name_file}{character. Name of the file to create, \code{gecevar} or \code{environ}.}
}
\description{
internal function.
}
