% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{declination}
\alias{declination}
\title{Declination}
\usage{
declination(jd)
}
\arguments{
\item{jd}{Julian Day.}
}
\value{
Declination in degrees and decimal fraction.
}
\description{
Computes the declination of the Sun for a given Julian Day.
}
\examples{

\dontrun{
declination(120)

## Plot daily changes in declination
jdays=c(1:365)
plot(declination(jdays), xlab="days of the year",
     ylab="declination", type="l")
}

}
\references{
\url{https://gml.noaa.gov/gmd/grad/solcalc/calcdetails.html}

Meeus, J. 1999.  \emph{Astronomical Algorithms}. Willmann-Bell, Richmond,
Virginia, USA.

Reda, I. and Andreas, A. 2003. \emph{Solar Position Algorithm for Solar
Radiation Applications}. 55 pp.; NREL Report No. TP-560-34302, Revised
January 2008.  \url{https://www.nrel.gov/docs/fy08osti/34302.pdf}
}
\author{
Javier G. Corripio \href{mailto:jgc@meteoexploration.com}{jgc@meteoexploration.com}
}
\keyword{internal}
