% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa_current.R
\name{get_chelsa_current}
\alias{get_chelsa_current}
\title{Create multilayer Tiff file with 107 variables from chelsa-climate.org}
\usage{
get_chelsa_current(
  extent,
  extent_latlon,
  EPSG,
  destination,
  resolution = 1000,
  rm_download = FALSE
)
}
\arguments{
\item{extent}{character. First output of \code{transform_shp_coutry_extent}.}

\item{extent_latlon}{character. Second output of \code{transform_shp_country_extent}}

\item{EPSG}{int. to consider for this country/area.}

\item{destination}{character. absolute path where to download files like \code{here()} output.}

\item{resolution}{int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.}

\item{rm_download}{boolean. If TRUE remove download files and folders. Keep only environ.tif in \code{data_raw} folder, default is FALSE.}
}
\value{
character. absolute path to climate_chelsa.tif.
}
\description{
Gives the average values on the data recovered between 1981 and 2010.
Monthly variables are average temperatures, min temperatures, max temperatures, precipitation, potential evapotranspiration (with Penman formula and with Thornthwaite formula),
and total cloud cover.
Others variables are climatic water deficit (with Penman and Thornthwaite), number of dry month (with Penman and Thornthwaite) and 19 bio variables (more information in chelsa documentation).
}
\details{
\code{resolution} need to be carefully choosen because if Tiff file is too big, R can crash.

Unit of each climatic variable :\tabular{ll}{
   Name \tab Unit \cr
   Temperature average (tas) \tab °C x 10 \cr
   Temperature min (tasmin) \tab °C x 10 \cr
   Temperature max (tasmax) \tab °C x 10 \cr
   Precipitation \tab kg.m^{-1} \cr
   Cloud area fraction (clt) \tab \% \cr
   PET Penman \tab kg.m^{-1} \cr
   PET Thornthwaite \tab kg.m^{-1} \cr
   Climatic water deficit (Penman) \tab kg.m^{-1} \cr
   Climatic dater deficit (Thornthwaite) \tab kg.m^{-1} \cr
   Number of dry months (Penman) \tab month \cr
   Number of dry months (Thornthwaite) \tab month \cr
   bio1 \tab °C x 10 \cr
   bio2 \tab °C x 10 \cr
   bio3 \tab °C x 10 \cr
   bio4 \tab °C x 10 \cr
   bio5 \tab °C x 10 \cr
   bio6 \tab °C x 10 \cr
   bio7 \tab °C x 10 \cr
   bio8 \tab °C x 10 \cr
   bio9 \tab °C x 10 \cr
   bio10 \tab °C x 10 \cr
   bio11 \tab °C x 10 \cr
   bio12 \tab kg.m^{-2}.year^{-1} \cr
   bio13 \tab kg.m^{-2}.month^{-1} \cr
   bio14 \tab kg.m^{-2}.month^{-1} \cr
   bio15 \tab kg.m^{-2} \cr
   bio16 \tab kg.m^{-2}.month^{-1} \cr
   bio17 \tab kg.m^{-2}.month^{-1} \cr
   bio18 \tab kg.m^{-2}.month^{-1} \cr
   bio19 \tab kg.m^{-2}.month^{-1} \cr
}
}
