% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa_current.R
\name{get_chelsa_current}
\alias{get_chelsa_current}
\title{Create multilayer Tiff file with 107 variables from chelsa-climate.org}
\usage{
get_chelsa_current(
  extent_latlon,
  extent_proj,
  EPSG_proj,
  destination,
  resol = 1000,
  rm_download = FALSE
)
}
\arguments{
\item{extent_latlon}{num vector. Must be in the form c(lonmin, latmin,
lonmax, latmax). This extent is used to download climatic data
using GDAL function \code{gdal_translate} and properties of COG
files. This extent can be obtained with the \code{get_aoi_extent()}
function.}

\item{extent_proj}{num vector. Must be in the form c(xmin, ymin, xmax,
ymax). This extent is used for the output raster file. This
extent can be obtained with the \code{get_aoi_extent()} function.}

\item{EPSG_proj}{int. EPSG code used to define the coordinate
reference system to project the output raster file.}

\item{destination}{character. Directory path for outputs.}

\item{resol}{int. Resolution. If in meters, recommended
resolutions are 250m, 500m, 1km, 2km or 5km. The resolution needs
to be carefully chosen. If set too small (e.g. < 250m), raster
file will be too big to fit in memory and R will crash. Default
is 1km.}

\item{rm_download}{boolean. If TRUE, remove downloaded files and
folders except \code{current_chelsa.tif} file in the \code{data_raw}
folder. Default is FALSE.}
}
\value{
character. The absolute path to the \code{current_chelsa.tif} file.
}
\description{
Gives the average values on the data recovered between
1981 and 2010. Monthly variables are average temperatures, min
temperatures, max temperatures, precipitation, potential
evapotranspiration (with Penman formula and with Thornthwaite
formula), and total cloud cover. Others variables are climatic
water deficit (with Penman and Thornthwaite), number of dry month
(with Penman and Thornthwaite) and 19 bio variables (more
information in chelsa documentation).
}
\details{
Unit of each climatic variable :\tabular{ll}{
   Name \tab Unit \cr
   Temperature average (tas) \tab °C x 10 \cr
   Temperature min (tasmin) \tab °C x 10 \cr
   Temperature max (tasmax) \tab °C x 10 \cr
   Precipitation \tab kg.m^{-2} \cr
   Cloud area fraction (clt) \tab \% \cr
   PET Penman \tab kg.m^{-2} \cr
   PET Thornthwaite \tab kg.m^{-2} \cr
   Climatic water deficit (Penman) \tab kg.m^{-2} \cr
   Climatic dater deficit (Thornthwaite) \tab kg.m^{-2} \cr
   Number of dry months (Penman) \tab month \cr
   Number of dry months (Thornthwaite) \tab month \cr
   bio1 \tab °C x 10 \cr
   bio2 \tab °C x 10 \cr
   bio3 \tab °C x 10 \cr
   bio4 \tab °C x 10 \cr
   bio5 \tab °C x 10 \cr
   bio6 \tab °C x 10 \cr
   bio7 \tab °C x 10 \cr
   bio8 \tab °C x 10 \cr
   bio9 \tab °C x 10 \cr
   bio10 \tab °C x 10 \cr
   bio11 \tab °C x 10 \cr
   bio12 \tab kg.m^{-2}.year^{-1} \cr
   bio13 \tab kg.m^{-2}.month^{-1} \cr
   bio14 \tab kg.m^{-2}.month^{-1} \cr
   bio15 \tab kg.m^{-2} \cr
   bio16 \tab kg.m^{-2}.month^{-1} \cr
   bio17 \tab kg.m^{-2}.month^{-1} \cr
   bio18 \tab kg.m^{-2}.month^{-1} \cr
   bio19 \tab kg.m^{-2}.month^{-1} \cr
}
}
