% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_variables.R
\name{get_env_variables}
\alias{get_env_variables}
\title{Create multilayer Tiff file with 11 environmental variables}
\usage{
get_env_variables(extent_latlon, extent, EPSG, country_name, destination)
}
\arguments{
\item{extent_latlon}{int vector. in this order c(lon_min, lat_min, lon_max, lat_max).}

\item{extent}{character. First output of \code{transform_shp_country_extent} function.}

\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{character. country name (in english) to considers.}

\item{destination}{character. absolute path where to download files like \code{here()} output.}
}
\value{
character. absolute path to environ.tif file.
}
\description{
Variables are type of soil, elevation, slope, aspect, roughness, solar radiation, distance to sea,
protected areas, distance to roads, distance to cities and town, distance to rivers & lake.
}
