% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_variables.R
\name{get_env_variables}
\alias{get_env_variables}
\title{Create multilayer Tiff file with 11 environmental variables}
\usage{
get_env_variables(
  extent_latlon,
  extent,
  EPSG,
  country_name = NULL,
  destination,
  resolution = 1000,
  rm_download = FALSE
)
}
\arguments{
\item{extent_latlon}{int vector. in this order c(lon_min, lat_min, lon_max, lat_max).}

\item{extent}{character. First output of \code{transform_shp_country_extent} function.}

\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{(optional) character. country name (in english) which be use to collect protected areas. This country must be available in \verb{https://www.protectedplanet.net/en/thematic-areas/wdpa?tab=WDPA}, default is NULL.}

\item{destination}{character. absolute path where to download files like \code{here()} output.}

\item{resolution}{int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.}

\item{rm_download}{boolean. If TRUE remove download files and folders. Keep only environ.tif in \code{data_raw} folder, default is FALSE.}
}
\value{
character. absolute path to environ.tif file.
}
\description{
Variables are type of soil, elevation, slope, aspect, roughness, solar radiation, distance to sea,
protected areas, distance to roads, distance to cities and town, distance to rivers & lake.
}
\details{
\code{resolution} need to be carefully choosen because if Tiff file is too big, R can crash.

Unit of each environ variable :\tabular{ll}{
   Name \tab Unit \cr
   Elevation \tab m \cr
   Aspect \tab degrees \cr
   Roughness \tab m \cr
   Slope \tab degrees \cr
   Solar radiation \tab Wh.m^{-2}.day \cr
   Soilgrids \tab category \cr
   Distance sea \tab m \cr
   Distance road \tab m \cr
   Distance place \tab m \cr
   Distance watering place \tab m \cr
   Protected Area (WDPA) \tab category \cr
}
}
