% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Create a multilayer Tiff file with environ & climat variables.}
\usage{
merge_files(environ_path, climat_path, destination)
}
\arguments{
\item{environ_path}{character. absolute path to environ.tif file, output of \code{get_env_variables}.}

\item{climat_path}{character. absolute path to current_chelsa.tif.tif file, output of \code{get_chelsa_variables}}

\item{destination}{character. absolute path where to download files like \code{here()} output.}
}
\value{
character. absolute path to gecevar.tif.
}
\description{
Use output of \code{get_env_variables} and \code{get_chelsa_variables} to create an unique file with more than 100 variables.
}
