% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_shp_country_extent.R
\name{transform_shp_country_extent}
\alias{transform_shp_country_extent}
\title{Create extent from country name or a shp file.}
\usage{
transform_shp_country_extent(
  EPSG,
  country_name = NULL,
  shapefile_path = NULL,
  extent_short = NULL
)
}
\arguments{
\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{character. English country name. Check \code{details} for more information, default is NULL.}

\item{shapefile_path}{character. Path to a .shp file, default is NULL.}

\item{extent_short}{int vector. length 4, with this order c(xmin, ymin, xmax, ymax), default is NULL.}
}
\value{
character vector. with extent of the area in one character, other are latlon coord of area
}
\description{
Create extent from country name or a shp file.
}
\details{
\code{country_name} must be available on \verb{https://www.naturalearthdata.com/} website.
}
