% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_shp_country_extent.R
\name{transform_shp_country_extent}
\alias{transform_shp_country_extent}
\title{Create extent & extent in latitude and longitude}
\usage{
transform_shp_country_extent(
  EPSG,
  country_name = NULL,
  shapefile_path = NULL,
  extent_short = NULL
)
}
\arguments{
\item{EPSG}{int. to consider for this country/area.}

\item{country_name}{character. English country name, default is NULL.}

\item{shapefile_path}{character. Path to a .shp file, default is NULL.}

\item{extent_short}{int vector. length 4, with this order c(xmin, ymin, xmax, ymax), default is NULL.}
}
\value{
character vector. with extent of the area in one character, other are latlon coord of area
}
\description{
With a country name, a shapefile or a extent and an EPSG value.
Succed to give extent in EPSG reproject and extent with default EPSG also knows as latitude and longitude coordinates.
}
